+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.1
+
+* changed the behavior of loop in `IntervalTerminableThread`