+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.1
 
 * changed the behavior of loop in `IntervalTerminableThread`
+* bugfix: fixed collector's logic